Added helper class for accessing and updating daily stats
