+++ Client/app/src/main/java/com/acme/snapgreen/data/StatUtil.kt
@@ -1,0 +1,64 @@
+package com.acme.snapgreen.data
+import android.content.Context
+import android.util.Log
+import io.realm.Realm
+import io.realm.kotlin.where
+import java.text.DateFormat
+import java.util.*
+
+/**
+ * A utility class to abstract the database component of updating daily usage.
+ * Allows caller to get today's DailyStatistic and update its values
+ */
+@Suppress("SpellCheckingInspection")
+class StatUtil private constructor() {
+
+    companion object {
+
+        /**
+         * Returns the DailyStatistic associated with today's date from Realm DB, or creates one if
+         * it doesn't exist.
+         */
+        fun getTodaysStats() : DailyStatistic {
+
+            // get the date and parse out the specific time (we just want day/month/year)
+            val dateStringList = DateFormat.getDateTimeInstance().format(Date()).split(",", " ")
+            val dateString = dateStringList[0] + " " + dateStringList[1] + " " + dateStringList[3]
+            val realm = Realm.getDefaultInstance()
+
+            // query database for today's date to see if we've already created a DailyStat for today
+            var stat = realm.where<DailyStatistic>().
+            contains("today",  dateString).findFirst()
+            realm.beginTransaction()
+
+            if(stat == null) {
+                // create entry for today's date if it does not exist
+                stat = DailyStatistic().apply {
+                    this.today = dateString
+                }
+                realm.copyToRealm(stat)
+                Log.i("Realm Database",
+                    "Created DailyStatistic for $dateString"
+                )
+            }
+            else {
+                Log.i("Realm Database",
+                    "Returning previously created DailyStatistic for $dateString")
+            }
+
+            return stat
+        }
+
+        /**
+         * Updates the DailyStatistic in the database
+         */
+        fun setTodaysStats(stats: DailyStatistic) {
+            val realm = Realm.getDefaultInstance()
+            realm.copyToRealmOrUpdate(stats)
+            realm.commitTransaction()
+            Log.i("Realm Database",
+                "Updating statistics associated with ${stats.today}")
+        }
+    }
+
+}
